Added new hook - UseFetchingOffer
